requirements.txt
asn1crypto==1.2.0
astroid==2.3.3
base58==1.0.3
bigchaindb-driver==0.6.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptoconditions==0.8.0
cryptography==2.3.1
dataclasses==0.6
Django==2.2.6
django-cors-headers==3.1.1
django-debug-toolbar==2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
dnspython==1.16.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pkg-resources==0.0.0
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.9.0
PyNaCl==1.1.2
pysha3==1.0.2
python-dateutil==2.8.1
python-rapidjson==0.6.3
python-rapidjson-schema==0.1.1
pytz==2019.2
requests==2.22.0
six==1.13.0
sqlparse==0.2.4
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
